servers/health-graph-mcp/pyproject.toml
[project]
name = "health-graph-mcp"
version = "0.1.0"
description = "Assertion-first health graph MCP server for genomics, labs, and coverage intelligence"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "asyncpg>=0.30.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.8.0",
    "stewardos-lib",
]

[tool.uv.sources]
stewardos-lib = { path = "../lib", editable = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
